pyproject.toml
[tool.poetry]
name = "workshop"
version = "0.1.0"
description = ""
authors = ["2004santiago <santiago.gomez.castro.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
pandas = "^2.2.2"
matplotlib = "^3.9.2"
numpy = "^2.1.0"
python-dotenv = "^1.0.1"
pg8000 = "^1.31.2"
sqlalchemy = "^2.0.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
